Adding more logging. (#12)
Logging "Done" so I know when an instance is done vs being killed or
encountering a bug.

Logging when a lock has been obtained since it doesn't log on every
iteration that it's still waiting for a lock.

Logging the new PID of do_toggled_on so I know the chain of events when
viewing logs.

Cleaning up README now that script is mostly implemented.