Closing inherited lock from gl-switch.
gl-switch uses `flock` to ensure a single instance of the switch script
runs when user toggles it. This gets inherited to all subprocesses with
interfered with this script's flock.

Releasing the inherited lock so multiple instances of this script can
run. Allows user to cancel "waiting for internet" loops when they toggle
the switch OFF.