Always release inherited locks (#13)
* Closing inherited lock from gl-switch.

gl-switch uses `flock` to ensure a single instance of the switch script
runs when user toggles it. This gets inherited to all subprocesses with
interfered with this script's flock.

Releasing the inherited lock so multiple instances of this script can
run. Allows user to cancel "waiting for internet" loops when they toggle
the switch OFF.

* todo

* todo

* info

* todo

* todo

* all debug in single instance func

* Cant close inherited, unlock instead

* grammar

* Always unlock inherited

Managing my own locks, no need for gl-switch or hotplug to apply
locking.

* revert logging changes